requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Environment dependencies
minigrid>=2.3.0
gymnasium>=0.29.0

# Data processing
datasets>=2.14.0
pandas>=2.0.0
h5py>=3.9.0

# Model dependencies (LLaDA2.0 / dFactory)
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# Evaluation
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
